Use the imageUpldFIFO from the rpimgdb: remove the local files which are not in the camera buffer and are in the uploaded images buffer

+++ rpimgdir.py
@@ -16,8 +16,8 @@
     You should have received a copy of the GNU General Public License along
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-    
-Implements the rpiImageDir class to manage the set of saved images by rpiCam    
+
+Implements the rpiImageDir class to manage the set of saved images by rpiCam
 """
 
 import os
@@ -36,25 +36,28 @@ class rpiImageDirClass(rpiBaseClass):
 	Implements the rpiImageDir class to manage the set of saved images by rpiCam
 	"""
 
-	def __init__(self, name, rpi_apscheduler, rpi_events, rpi_config, dbuff_rpififo=None):
-	
-		### Get the Dbx error event	
-		self._eventDbErr 	= rpi_events.eventErrList["DBXJob"] 
+	def __init__(self, name, rpi_apscheduler, rpi_events, rpi_config, cam_rpififo=None, upld_rpififo=None):
 
-		### Get the custom config parameters			
+		### Get the Dbx error event
+		self._eventDbErr 	= rpi_events.eventErrList["DBXJob"]
+
+		### Get the custom config parameters
 		self._config = rpi_config
-	
-		### Get FIFO buffer (deque)							
-		self._imageFIFO = dbuff_rpififo
-				
+
+		### Get FIFO buffer for images from the camera (deque)
+		self._imageFIFO = cam_rpififo
+
+		### Get FIFO buffer for the uploaded images (deque)
+		self._imageUpldFIFO = upld_rpififo
+
 		### Init base class
 		super().__init__(name, rpi_apscheduler, rpi_events)
 
 	def __repr__(self):
 		return "<%s (name=%s, rpi_apscheduler=%s, rpi_events=dict(), rpi_config=%s, dbuff_rpififo=%s)>" % (self.__class__.__name__, self.name, self._sched, self._config, self._imageFIFO)
-						
+
 	def __str__(self):
-		msg = super().__str__()	
+		msg = super().__str__()
 		return "%s::: config: %s, locdir: %s, image_names: %s, len(imagelist_ref): %d\n%s" % \
 				(self.name, self._config, self._locdir, self._image_names, len(self._imagelist_ref), msg)
 
@@ -65,11 +68,11 @@ class rpiImageDirClass(rpiBaseClass):
 
 	#
 	# Main interface methods
-	#		
-	
+	#
+
 	def jobRun(self):
-				
-								
+
+
 		### List all jpg files in the current local sub-folder
 		self._locdir = os.path.join(self._config['image_dir'], self._imageFIFO.crtSubDir)
 		self._image_names = os.path.join(self._locdir, self._imageFIFO.crtSubDir + '-*' + self._imageFIFO.camID + '.jpg')
@@ -78,49 +81,51 @@ class rpiImageDirClass(rpiBaseClass):
 			logging.debug("imagelist: %s .. %s" % (self.imagelist[0], self.imagelist[-1]))
 		else:
 			logging.debug("imagelist: empty. No %s found!" % self._image_names)
-		
-		### Run directory/file management only if no errors were detected when 
+
+		### Run directory/file management only if no errors were detected when
 		### updating to remote directory
 		if not self._eventDbErr.is_set():
 			# Process the new list only if it is changed and has at least max length
 			if ( not (self._imagelist_ref == self.imagelist) ) and \
 				len(self.imagelist) > self._config['list_size']:
-	
-				# Remove all the images not in the imageFIFO
+
+				# Remove all the local images, which are
+				# not in the camera buffer and are in the uploaded images buffer
 				try:
 					self._imageFIFO.acquireSemaphore()
-		
+
 					for img in self.imagelist:
-						if not img in self._imageFIFO:					
-							logging.info("Remove image: %s" % img)				
-							self._rmimg = subprocess.Popen("rm " + img, stderr=subprocess.PIPE, stdout=subprocess.PIPE, shell=True) 
+						if not img in self._imageFIFO and \
+							img in self._imageUpldFIFO:
+							logging.info("Remove image: %s" % img)
+							self._rmimg = subprocess.Popen("rm " + img, stderr=subprocess.PIPE, stdout=subprocess.PIPE, shell=True)
 							self._diroutput, self._direrrors = self._rmimg.communicate()
-			
-							
+
+
 				except OSError as e:
 					raise rpiBaseClassError("%s::: jobRun(): File %s could not be deleted!\n%s" % (self.name, img, e), ERRLEV2)
-				
+
 				except:
 					raise rpiBaseClassError("%s::: jobRun(): Unhandled Exception:\n%s" % (self.name, str(sys.exc_info())), ERRCRIT)
-				
-				finally:		
+
+				finally:
 					self._imageFIFO.releaseSemaphore()
-			
-			#raise rpiBaseClassError("%s::: jobRun(): Test crash!" % self.name, 4)	
-				
+
+			#raise rpiBaseClassError("%s::: jobRun(): Test crash!" % self.name, 4)
+
 			# Update status
 			self.statusUpdate = (self.name, len(self.imagelist))
-		
+
 			# Update image list in the current local sub-folder
 			self._imagelist_ref = sorted(glob.glob(self._image_names))
 			if len(self._imagelist_ref) > 0:
 				logging.debug("imagelist_ref: %s .. %s" % (self._imagelist_ref[0], self.imagelist[-1]))
 			else:
 				logging.debug("imagelist_ref: empty. No %s found!" % self._image_names)
-		
-				
+
+
 		else:
-			logging.info("eventDbErr is set!")							
+			logging.info("eventDbErr is set!")
 
 
 
@@ -131,16 +136,16 @@ class rpiImageDirClass(rpiBaseClass):
 
 		### Init reference img file list
 		self._locdir = os.path.join(self._config['image_dir'], self._imageFIFO.crtSubDir)
-		self._image_names = os.path.join(self._locdir, self._imageFIFO.crtSubDir + '-*' + self._imageFIFO.camID + '.jpg')		
+		self._image_names = os.path.join(self._locdir, self._imageFIFO.crtSubDir + '-*' + self._imageFIFO.camID + '.jpg')
 		self._imagelist_ref = sorted(glob.glob(self._image_names))
 
-		
+
 #	def endDayOAM(self):
 #		"""
 #		End-of-Day 0AM
-#		"""	
-		
+#		"""
+
 #	def endOAM(self):
 #		"""
 #		End OAM procedure.
-#		"""	
+#		"""
